[TSLO] Create AppHeaderCoordinator during Activity.onCreate()
Create the coordinator early before the first inflation pass so that the
temp tab strip drawable can draw into the app header.

Before v.s. After comparison with Canary:
https://screencast.googleplex.com/cast/NDk4ODgwOTI4MzM3MTAwOHxiMTdjOGNlOC1iYQ

Bug: 325351108
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5417874
Commit-Queue: Wenyu Fu <wenyufu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1282970}

+++ chrome/android/junit/src/org/chromium/chrome/browser/compositor/overlays/strip/StripLayoutHelperManagerTest.java
@@ -601,8 +601,8 @@ public class StripLayoutHelperManagerTest {
         activeLayoutHelper.setLastHoveredTabForTesting(mHoveredStripTab);
 
         // Update the paddings.
-        float leftPadding = 10f;
-        float rightPadding = 20f;
+        int leftPadding = 10;
+        int rightPadding = 20;
         int topPaddingPx = 5;
         mStripLayoutHelperManager.updateHorizontalPaddings(leftPadding, rightPadding);
         mStripLayoutHelperManager.onHeightChanged(TAB_STRIP_HEIGHT_PX + topPaddingPx);
@@ -624,8 +624,8 @@ public class StripLayoutHelperManagerTest {
                         eq(hoveredTabId),
                         anyInt(),
                         anyFloat(),
-                        eq(leftPadding),
-                        eq(rightPadding),
+                        eq((float) leftPadding),
+                        eq((float) rightPadding),
                         eq((float) topPaddingPx));
     }
 
@@ -948,8 +948,8 @@ public class StripLayoutHelperManagerTest {
     @Test
     public void testTouchEventsIgnoredOnPaddings() {
         // Update the size and paddings.
-        float leftPadding = 10f;
-        float rightPadding = 20f;
+        int leftPadding = 10;
+        int rightPadding = 20;
         mStripLayoutHelperManager.updateHorizontalPaddings(leftPadding, rightPadding);
         mStripLayoutHelperManager.onSizeChanged(
                 SCREEN_WIDTH, SCREEN_HEIGHT, VISIBLE_VIEWPORT_Y, ORIENTATION);
